Add sticky sidebars for Extraction, Notification, Entry page
- Add "Figures with Review Requested" tab on QA
- Use hash on QA tabs
- Fix navbar height
- Update toggle-ui
- Create a BasicItem component
- Move EntryForm to components/
- Create a Fragment for Figure
- Reuse preview component
