+++ src/Root/App/styles.css
@@ -33,6 +33,8 @@
     --width-separator-medium: 2px;
     --width-separator-thick: 4px;
 
+    --height-navbar: 58px;
+
     /* FONT SIZE */
     --font-size-extra-small: 11px;
     --font-size-small: 12px;
