+++ src/components/GeoInput/index.tsx
@@ -26,7 +26,7 @@ import {
 import { _cs, isDefined, isTruthyString } from '@togglecorp/fujs';
 import { mergeBbox } from '#utils/common';
 
-import { GeoLocationFormProps } from '#views/Entry/EntryForm/types';
+import { GeoLocationFormProps } from '#components/forms/EntryForm/types';
 import Loading from '#components/Loading';
 import {
     LookupQuery,
