+++ src/components/forms/EntryForm/DetailsInput/styles.css
