+++ src/components/forms/EntryForm/FigureInput/styles.css
@@ -36,7 +36,7 @@
         height: 300px;
     }
 
-    .action-button {
+    .action-buttons {
         display: flex;
         gap: var(--spacing-medium);
         justify-content: flex-end;
