+++ src/components/tables/EventsEntriesFiguresTable/NudeEventsTable/index.tsx
@@ -53,46 +53,46 @@ type EventFields = NonNullable<NonNullable<EventListQuery['eventList']>['results
 
 export const EVENT_LIST = gql`
     query EventList(
-    $ordering: String,
-    $page: Int,
-    $pageSize: Int,
-    $name: String,
-    $eventTypes:[String!],
-    $crisisByIds: [ID!],
-    $countries:[ID!],
-    $violenceSubTypes: [ID!],
-    $disasterSubTypes: [ID!],
-    $contextOfViolences: [ID!],
-    $osvSubTypeByIds: [ID!],
-    $glideNumbers: [String!],
-    $createdByIds: [ID!],
-    $startDate_Gte: Date,
-    $endDate_Lte: Date,
-    $qaRules: [String!],
-    $ignoreQa: Boolean,
-    $reviewStatus: [String!],
-    $assignees: [ID!],
+        $ordering: String,
+        $page: Int,
+        $pageSize: Int,
+        $name: String,
+        $eventTypes:[String!],
+        $crisisByIds: [ID!],
+        $countries:[ID!],
+        $violenceSubTypes: [ID!],
+        $disasterSubTypes: [ID!],
+        $contextOfViolences: [ID!],
+        $osvSubTypeByIds: [ID!],
+        $glideNumbers: [String!],
+        $createdByIds: [ID!],
+        $startDate_Gte: Date,
+        $endDate_Lte: Date,
+        $qaRules: [String!],
+        $ignoreQa: Boolean,
+        $reviewStatus: [String!],
+        $assignees: [ID!],
     ) {
         eventList(
-        ordering: $ordering,
-        page: $page,
-        pageSize: $pageSize,
-        contextOfViolences: $contextOfViolences,
-        name: $name,
-        eventTypes:$eventTypes,
-        crisisByIds: $crisisByIds,
-        countries: $countries,
-        violenceSubTypes: $violenceSubTypes,
-        disasterSubTypes: $disasterSubTypes,
-        createdByIds: $createdByIds,
-        startDate_Gte: $startDate_Gte,
-        endDate_Lte: $endDate_Lte,
-        qaRules: $qaRules,
-        ignoreQa: $ignoreQa,
-        osvSubTypeByIds: $osvSubTypeByIds,
-        glideNumbers: $glideNumbers,
-        reviewStatus: $reviewStatus,
-        assignees: $assignees,
+            ordering: $ordering,
+            page: $page,
+            pageSize: $pageSize,
+            contextOfViolences: $contextOfViolences,
+            name: $name,
+            eventTypes:$eventTypes,
+            crisisByIds: $crisisByIds,
+            countries: $countries,
+            violenceSubTypes: $violenceSubTypes,
+            disasterSubTypes: $disasterSubTypes,
+            createdByIds: $createdByIds,
+            startDate_Gte: $startDate_Gte,
+            endDate_Lte: $endDate_Lte,
+            qaRules: $qaRules,
+            ignoreQa: $ignoreQa,
+            osvSubTypeByIds: $osvSubTypeByIds,
+            glideNumbers: $glideNumbers,
+            reviewStatus: $reviewStatus,
+            assignees: $assignees,
         ) {
             totalCount
             pageSize
