+++ src/views/Entry/styles.css
@@ -1,70 +1,44 @@
 .entry {
     display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-    overflow: auto;
+    width: 100%;
     gap: var(--spacing-large);
 
-    .header {
+    .side-content {
+        display: flex;
+        flex-direction: column;
         flex-shrink: 0;
+        width: calc(160px + 30vw);
+        gap: var(--spacing-large);
 
-        .portal-container {
-            display: flex;
-            align-items: center;
-            gap: var(--spacing-small);
+        .sticky-container {
+            position: sticky;
+            top: var(--spacing-large);
+            height: calc(100vh - 2 * var(--spacing-large) - var(--height-navbar));
         }
     }
 
-    .content {
+    .main-content {
         display: flex;
+        flex-direction: column;
         flex-grow: 1;
-        overflow: auto;
+        overflow-x: auto;
         gap: var(--spacing-large);
+
+        .header {
+            flex-shrink: 0;
+
+            .portal-container {
+                display: flex;
+                align-items: center;
+                gap: var(--spacing-small);
+            }
+        }
 
         .entry-form {
             flex-grow: 1;
             border: var(--width-separator-thin) solid var(--color-separator-light);
             border-radius: var(--radius-card);
             overflow: auto;
-        }
-
-        .aside {
-            display: flex;
-            flex-direction: column;
-            flex-grow: 0;
-            flex-shrink: 0;
-            border: var(--width-separator-thin) solid var(--color-separator-light);
-            border-radius: var(--radius-card);
-            background-color: var(--color-foreground);
-            width: calc(160px + 30vw);
-
-            .tab-list {
-                display: flex;
-                flex-shrink: 0;
-
-                >* {
-                    flex-basis: 0;
-                    flex-grow: 1;
-
-                    .status-icon {
-                        margin: 0 var(--spacing-small)
-                    }
-                }
-
-                .preview-failed {
-                    color: var(--color-danger);
-                }
-            }
-
-            .preview-container {
-                display: flex;
-                flex-direction: column;
-                flex-grow: 1;
-
-                .preview {
-                    flex-grow: 1;
-                }
-            }
         }
     }
 }
