+++ src/views/Extraction/SavedFiltersList/FilterItem/styles.css
@@ -1,22 +1,3 @@
-.item-row {
-    display: flex;
-    align-items: center;
-    padding: var(--spacing-small) var(--spacing-medium);
-
-    .name {
-        flex: 1;
-    }
-
-    .action-buttons {
-        display: flex;
-        justify-content: space-between;
-        visibility: hidden;
-    }
-
-    &:hover {
-        .action-buttons {
-            visibility: visible;
-        }
-    }
+.name {
+    flex-grow: 1;
 }
-
