+++ src/views/Extraction/SavedFiltersList/styles.css
@@ -5,9 +5,5 @@
         display: flex;
         position: relative;
         flex-direction: column;
-
-        .selected-query {
-            background: var(--color-background);
-        }
     }
 }
