+++ src/views/Extraction/styles.css
@@ -21,6 +21,12 @@
         min-height: 400px;
     }
 
+    .sticky-container {
+        position: sticky;
+        top: var(--spacing-large);
+        min-height: 400px;
+    }
+
     .main-content {
         display: flex;
         flex-direction: column;
