+++ src/views/Notifiations/NotificationCategory/styles.css
@@ -3,11 +3,7 @@
     flex-direction: column;
 
     .category-name {
-        padding: var(--tui-spacing-medium);
+        padding: var(--tui-spacing-small);
         text-decoration: none;
-
-        &:hover {
-            background-color: var(--tui-color-background-hover-light)
-        }
     }
 }
